requirements.txt
anyio==4.10.0
asgiref==3.9.1
branca==0.8.1
build==1.3.0
CacheControl==0.14.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
cryptography==45.0.6
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dulwich==0.22.8
factory_boy==3.3.3
Faker==37.5.3
fastjsonschema==2.21.2
filelock==3.18.0
findpython==0.6.3
folium==0.20.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
keyring==25.6.0
lxml==6.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
msgpack==1.1.1
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pbs-installer==2025.8.14
pipenv==2025.0.4
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pypi==2.1
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-django==4.11.1
python-docx==1.2.0
python3-openid==3.2.0
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tomlkit==0.13.3
trove-classifiers==2025.8.6.13
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
xyzservices==2025.4.0
zstandard==0.24.0
